GlobalMessageBase: 	- removed added to factory bool as it is not used
LargePacketTransferManager:
	- changes the way the packet id is regiestered with class factory
	to be dynamic

SimStateSyncProcessor:
	- initalised m_bIsFullStateSynced to false on startup
	- registered required data packets with class factory
	- initalised m_dtmRequestTimeOut and m_dtmTimeOfNextTimeout on startup
	- reset data segment tracking arrays on hash change
	- set full state sync to false when hash changes
	- added the segment index to the segment hash to avoid matching
	hashes for duplicate segments
	- set full state sync to true when all segments recieved
	- fixed bug with getting next segment index from enumerator
	- fixed bug with segment time out not being calculated correctly
	- added funciton for getting list of requested sim state times
	- added function for setting sim state data for a peer

	- added start function for SimStateSyncConnectionProcessor to
	initalise all start values
	- changes the way the total sim state hash is created to be a littel
	more efficient

BaseNetworkTestSetup:
	-first pass of simple network setup to help build unit tests
	of packet processors

SimStateSyncUnitTest:
	- first pass testing framework for syncing sim data
	accross peers

TimePacketProcessor:
	- changed the way time data packets are registered with class
	factory to be dynamic

Serialization:
	- fixed bug with long array not serializing data correctly

+++ Assets/Code/Networking/Serializer/Serialization.cs
@@ -204,7 +204,7 @@ namespace Networking
 
             for (int i = 0; i < iItems; i++)
             {
-                Byte value = 0;
+                Int32 value = 0;
                 Serialize(rbsStream, ref value);
                 Output[i] = value;
             }
@@ -296,7 +296,7 @@ namespace Networking
 
             for (int i = 0; i < iItems; i++)
             {
-                Byte value = 0;
+                Int64 value = 0;
                 Serialize(rbsStream, ref value);
                 Output[i] = value;
             }
